+++ zad1/src/main/java/tul/ftims/cps/App.java
@@ -6,8 +6,8 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 import lombok.Data;
-import tul.ftims.cps.model.SignalManager;
-import tul.ftims.cps.model.TriangularSignal;
+import tul.ftims.cps.model.manager.SignalManager;
+import tul.ftims.cps.model.signals.TriangularSignal;
 
 
 /**
@@ -36,33 +36,33 @@ public class App extends Application {
 
 //        RectanguralSignal signal = new RectanguralSignal(5, 0, 100, 1, 0.5);
 //        System.out.println(signal.getSamples().toString());
-        //SinusoidalSignal signal = new SinusoidalSignal(5.0, 0.0, 10.0, 10.0,2.0);
-        //SinusoidalSignalHalfErected signal = new SinusoidalSignalHalfErected(5.0, 0.0, 10.0, 10.0,2.0);
-        //UnitJump signal = new UnitJump(5.0, 0.0, 10.0, 10.0, 2.0);
-        //UnitPulse signal = new UnitPulse(5.0, 0.0, 10.0, 10.0, 2.0);
-        //ImpulsiveNoise signal = new ImpulsiveNoise(5.0, 0.0, 10.0, 10.0);
-        //RectanguralSignal signal = new RectanguralSignal(5.0, 0.0, 10.0, 10.0, 2.0, 0.5);
-        //SymmetricalRectangularSignal signal = new SymmetricalRectangularSignal(5.0, 0.0, 10.0, 10.0, 2.0, 0.5);
-        TriangularSignal signal = new TriangularSignal(5.0, 0.0, 10.0, 10.0, 2.0, 0.5);
-
-
-        System.out.println("aaa");
-        System.out.println(signal.getMediumValue());
-        System.out.println(signal.getAbsoluteMediumValue());
-        System.out.println(signal.getMediumPower());
-        System.out.println(signal.getVariance());
-        System.out.println(signal.getEffectiveValue());
-        System.out.println(signal.getSamples());
-        System.out.println(signal.getSamples().size());
-
-        for (double key: signal.getSamples().keySet()){
-            System.out.print(key +" ");
-        }
-        System.out.println("");
-        System.out.println("aaa");
-
-        for (double key: signal.getSamples().keySet()){
-            System.out.print(signal.getSamples().get(key) +" ");
-        }
+//        SinusoidalSignal signal = new SinusoidalSignal(5.0, 0.0, 10.0, 10.0,2.0);
+//        SinusoidalSignalHalfErected signal = new SinusoidalSignalHalfErected(5.0, 0.0, 10.0, 10.0,2.0);
+//        UnitJump signal = new UnitJump(5.0, 0.0, 10.0, 10.0, 2.0);
+//        UnitPulse signal = new UnitPulse(5.0, 0.0, 10.0, 10.0, 2.0);
+//        ImpulsiveNoise signal = new ImpulsiveNoise(5.0, 0.0, 10.0, 10.0);
+//        RectanguralSignal signal = new RectanguralSignal(5.0, 0.0, 10.0, 10.0, 2.0, 0.5);
+//        SymmetricalRectangularSignal signal = new SymmetricalRectangularSignal(5.0, 0.0, 10.0, 10.0, 2.0, 0.5);
+//        TriangularSignal signal = new TriangularSignal(5.0, 0.0, 10.0, 10.0, 2.0, 0.5);
+//
+//
+//        System.out.println("aaa");
+//        System.out.println(signal.getMediumValue());
+//        System.out.println(signal.getAbsoluteMediumValue());
+//        System.out.println(signal.getMediumPower());
+//        System.out.println(signal.getVariance());
+//        System.out.println(signal.getEffectiveValue());
+//        System.out.println(signal.getSamples());
+//        System.out.println(signal.getSamples().size());
+//
+//        for (double key : signal.getSamples().keySet()) {
+//            System.out.print(key + " ");
+//        }
+//        System.out.println("");
+//        System.out.println("aaa");
+//
+//        for (double key : signal.getSamples().keySet()) {
+//            System.out.print(signal.getSamples().get(key) + " ");
+//        }
     }
 }
