+++ zad1/src/main/java/tul/ftims/cps/controller/MainController.java
@@ -6,10 +6,11 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.ListView;
 import javafx.scene.control.MenuItem;
+import javafx.scene.control.TextField;
 import javafx.scene.control.TitledPane;
 import javafx.stage.Stage;
 import tul.ftims.cps.App;
-import tul.ftims.cps.model.Signal;
+import tul.ftims.cps.model.manager.Signal;
 
 import java.io.IOException;
 
@@ -72,6 +73,39 @@ public class MainController {
     private TitledPane P_Statistics;
 
     @FXML
+    private TextField P_Amplitude;
+
+    @FXML
+    private TextField P_StartTime;
+
+    @FXML
+    private TextField P_Duration;
+
+    @FXML
+    private TextField P_SampleFrequency;
+
+    @FXML
+    private TextField P_BasicPeriod;
+
+    @FXML
+    private TextField P_FillFactor;
+
+    @FXML
+    private TextField P_AverageValue;
+
+    @FXML
+    private TextField P_MeanAbsolute;
+
+    @FXML
+    private TextField P_AverageStrength;
+
+    @FXML
+    private TextField P_Variance;
+
+    @FXML
+    private TextField P_EffectiveValue;
+
+    @FXML
     void initialize() {
 
         assert MB_F_NewProject != null : "fx:id=\"MB_F_NewProject\" was not injected: check your FXML file 'Main.fxml'.";
@@ -92,6 +126,18 @@ public class MainController {
         assert P_ListOfSignals != null : "fx:id=\"P_ListOfSignals\" was not injected: check your FXML file 'Main.fxml'.";
         assert P_ListOfSignals_ListView != null : "fx:id=\"P_ListOfSignals_ListView\" was not injected: check your FXML file 'Main.fxml'.";
         assert P_Statistics != null : "fx:id=\"P_Statistics\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_Amplitude != null : "fx:id=\"P_Amplitude\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_StartTime != null : "fx:id=\"P_StartTime\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_Duration != null : "fx:id=\"P_Duration\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_SampleFrequency != null : "fx:id=\"P_SampleFrequency\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_BasicPeriod != null : "fx:id=\"P_BasicPeriod\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_FillFactor != null : "fx:id=\"P_FillFactor\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_AverageValue != null : "fx:id=\"P_AverageValue\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_MeanAbsolute != null : "fx:id=\"P_MeanAbsolute\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_AverageStrength != null : "fx:id=\"P_AverageStrength\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_Variance != null : "fx:id=\"P_Variance\" was not injected: check your FXML file 'Main.fxml'.";
+        assert P_EffectiveValue != null : "fx:id=\"P_EffectiveValue\" was not injected: check your FXML file 'Main.fxml'.";
+
 
         MB_S_NewSignal.setOnAction(event -> {
             try {
@@ -108,7 +154,9 @@ public class MainController {
 
         P_ListOfSignals_ListView.setItems(signalManager.getRepository());
 
-        
+//        P_ListOfSignals_ListView.getSelectionModel().selectedItemProperty().addListener((InvalidationListener) (observable, oldValue, newValue) -> P_Amplitude.textProperty().setValue());
+
+//        P_Amplitude.textProperty().setValue(P_ListOfSignals_ListView.getSelectionModel().getSelectedItem().getAmplitude().toString()==null?);
 
     }
 }
