+++ zad1/src/main/java/tul/ftims/cps/model/manager/SignalCategory.java
@@ -1,4 +1,4 @@
-package tul.ftims.cps.model;
+package tul.ftims.cps.model.manager;
 
 public enum SignalCategory {
     DISCREET, CONTINUOUS
