+++ zad1/src/main/java/tul/ftims/cps/model/manager/SignalType.java
@@ -1,4 +1,4 @@
-package tul.ftims.cps.model;
+package tul.ftims.cps.model.manager;
 
 public enum SignalType {
