+++ zad1/src/main/java/tul/ftims/cps/model/operations/OperationType.java
@@ -1,4 +1,4 @@
-package tul.ftims.cps.model;
+package tul.ftims.cps.model.operations;
 
 public enum OperationType {
     ADD, SUBTRACT, MULTIPLY, DIVIDE
