+++ zad1/src/main/java/tul/ftims/cps/model/signals/RectanguralSignal.java
@@ -1,33 +1,35 @@
-package tul.ftims.cps.model;
+package tul.ftims.cps.model.signals;
+
+import tul.ftims.cps.model.manager.Signal;
+import tul.ftims.cps.model.manager.SignalCategory;
 
 import java.util.Map;
-import java.util.Random;
 
-public class GaussianNoise extends Signal {
+public class RectanguralSignal extends Signal {
 
-    public GaussianNoise(Double amplitude, Double startTime, Double duration, Double samplingFrequency) {
-        super(amplitude, startTime, duration, samplingFrequency, SignalCategory.CONTINUOUS);
+    Double kw; //współczynnik czasu trwania wartości maksymalnej do okresu
+
+    public RectanguralSignal(Double amplitude, Double startTime, Double duration, Double samplingFrequency, Double T, Double kw) {
+        super(amplitude, startTime, duration, samplingFrequency, T, SignalCategory.CONTINUOUS);
+        this.kw = kw;
         this.generate(getSamples());
         this.calculateValues();
     }
 
-    public GaussianNoise(double amplitude, double startTime, double duration, double samplingFrequency) {
-        super(Double.valueOf(amplitude), Double.valueOf(startTime), Double.valueOf(duration), Double.valueOf(samplingFrequency), SignalCategory.CONTINUOUS);
+    public RectanguralSignal(double amplitude, double startTime, double duration, double samplingFrequency, double T, double kw) {
+        super(Double.valueOf(amplitude), Double.valueOf(startTime), Double.valueOf(duration), Double.valueOf(samplingFrequency), Double.valueOf(T), SignalCategory.CONTINUOUS);
+        this.kw = kw;
         this.generate(getSamples());
         this.calculateValues();
     }
-
-    // Funkcja gęstości rozkładu zmiennej losowej
-    public Double probabilityDensityFunction(Double x) {
-        return (1 / Math.sqrt(2 * Math.PI)) * Math.exp(-(Math.pow(x.doubleValue(), 2)) / 2);
-    }
-
-    Double t2 = this.getStartTime() + this.getDuration(); // (t1 + d)
-    double minimalAmplitude = -Math.abs(this.getAmplitude().doubleValue()); // min(A)
-    double maximumAmplitude = Math.abs(this.getAmplitude().doubleValue()); // max(A)
 
     public double func(double t1) {
-        return probabilityDensityFunction(minimalAmplitude + new Random().nextDouble() * (maximumAmplitude - minimalAmplitude));
+        double result;
+        double k = Math.floor((t1 - this.getStartTime()) / this.getT());
+        if (k > (t1 - 0.5 * this.getT()) / this.getT())
+            result = this.getAmplitude();
+        else result = 0;
+        return result;
     }
 
     public double funcAbs(double t1) {
@@ -42,8 +44,9 @@ public class GaussianNoise extends Signal {
         return Math.pow(func(t1) - getMediumValue(), 2);
     }
 
+    //nie działa!!! //jednak działa!!
     public void generate(Map<Double, Double> samples) {
-
+        Double t2 = this.getStartTime() + this.getDuration(); // (t1 + d)
         for (Double t1 = this.getStartTime(); t1.compareTo(t2) < 0; t1 += 1 / this.getSamplingFrequency()) {
             samples.put(t1, func(t1));
         }
@@ -57,4 +60,6 @@ public class GaussianNoise extends Signal {
         this.setVariance(MediumValueORAbsolutMediumValueORMediumPowerORVarianceC(this.getStartTime(), t2, this::funcVar));
         this.setEffectiveValue(EffectiveValueC(this.getStartTime(), t2, this::funcPow));
     }
+
 }
+
