+++ zad1/src/main/java/tul/ftims/cps/model/signals/SinusoidalSignal.java
@@ -1,25 +1,28 @@
-package tul.ftims.cps.model;
+package tul.ftims.cps.model.signals;
+
+import lombok.Data;
+import tul.ftims.cps.model.manager.Signal;
+import tul.ftims.cps.model.manager.SignalCategory;
 
 import java.util.Map;
 
-public class SinusoidalSignalHalfErected extends Signal {
+@Data
+public class SinusoidalSignal extends Signal {
 
-    public SinusoidalSignalHalfErected(Double amplitude, Double startTime, Double duration, Double samplingFrequency, Double T) {
+    public SinusoidalSignal(Double amplitude, Double startTime, Double duration, Double samplingFrequency, Double T) {
         super(amplitude, startTime, duration, samplingFrequency, T, SignalCategory.CONTINUOUS);
         this.generate(getSamples());
         this.calculateValues();
     }
 
-    public SinusoidalSignalHalfErected(double amplitude, double startTime, double duration, double samplingFrequency, double T) {
+    public SinusoidalSignal(double amplitude, double startTime, double duration, double samplingFrequency, double T) {
         super(Double.valueOf(amplitude), Double.valueOf(startTime), Double.valueOf(duration), Double.valueOf(samplingFrequency), Double.valueOf(T), SignalCategory.CONTINUOUS);
         this.generate(getSamples());
         this.calculateValues();
     }
 
     public double func(double t1) {
-        return 0.5 * this.getAmplitude() *
-                (Math.sin(((2 * Math.PI) / this.getT()) * ((t1 - this.getStartTime())))
-                        + Math.abs(Math.sin(((2 * Math.PI) / this.getT()) * ((t1 - this.getStartTime())))));
+        return this.getAmplitude() * Math.sin(((2 * Math.PI) / this.getT()) * ((t1 - this.getStartTime())));
     }
 
     public double funcAbs(double t1) {
@@ -37,7 +40,7 @@ public class SinusoidalSignalHalfErected extends Signal {
 
     public void generate(Map<Double, Double> samples) {
         Double t2 = this.getStartTime() + this.getDuration(); // (t1 + d)
-        for (Double t1 = this.getStartTime(); t1.compareTo(t2) < 0; t1 += 1 / this.getT()) {
+        for (Double t1 = this.getStartTime(); t1.compareTo(t2) < 0; t1 += (1 / this.getSamplingFrequency())) {
             samples.put(t1, func(t1));
         }
     }
@@ -50,4 +53,5 @@ public class SinusoidalSignalHalfErected extends Signal {
         this.setVariance(MediumValueORAbsolutMediumValueORMediumPowerORVarianceC(this.getStartTime(), t2, this::funcVar));
         this.setEffectiveValue(EffectiveValueC(this.getStartTime(), t2, this::funcPow));
     }
+
 }
