+++ zad1/src/main/java/tul/ftims/cps/model/signals/SinusoidalSignalErected.java
@@ -1,33 +1,25 @@
-package tul.ftims.cps.model;
+package tul.ftims.cps.model.signals;
+
+import tul.ftims.cps.model.manager.Signal;
+import tul.ftims.cps.model.manager.SignalCategory;
 
 import java.util.Map;
 
-public class TriangularSignal extends Signal {
-
-    Double kw; //współczynnik czasu trwania wartości maksymalnej do okresu
-
-    public TriangularSignal(Double amplitude, Double startTime, Double duration, Double samplingFrequency, Double T, Double kw) {
+public class SinusoidalSignalErected extends Signal {
+    public SinusoidalSignalErected(Double amplitude, Double startTime, Double duration, Double samplingFrequency, Double T) {
         super(amplitude, startTime, duration, samplingFrequency, T, SignalCategory.CONTINUOUS);
-        this.kw = kw;
         this.generate(getSamples());
         this.calculateValues();
     }
 
-    public TriangularSignal(double amplitude, double startTime, double duration, double samplingFrequency, double T, double kw) {
+    public SinusoidalSignalErected(double amplitude, double startTime, double duration, double samplingFrequency, double T) {
         super(Double.valueOf(amplitude), Double.valueOf(startTime), Double.valueOf(duration), Double.valueOf(samplingFrequency), Double.valueOf(T), SignalCategory.CONTINUOUS);
-        this.kw = kw;
         this.generate(getSamples());
         this.calculateValues();
     }
 
     public double func(double t1) {
-        double result;
-        double k = Math.floor((t1 - this.getStartTime()) / this.getT());
-        if (k > (t1 - 0.5 * this.getT()) / this.getT())
-            result = (this.getAmplitude() / (kw * this.getT())) * (t1 - k * this.getT() - this.getStartTime());
-        else
-            result = ((0.0 - this.getAmplitude()) / (this.getT() * (1 - kw))) * (t1 - k * this.getT() - this.getStartTime()) + this.getAmplitude() / (1 - kw);
-        return result;
+        return this.getAmplitude() * Math.abs(Math.sin(((2 * Math.PI) / this.getT()) * ((t1 - this.getStartTime()))));
     }
 
     public double funcAbs(double t1) {
@@ -39,6 +31,7 @@ public class TriangularSignal extends Signal {
     }
 
     public double funcVar(double t1) {
+        Double t2 = this.getStartTime() + this.getDuration();
         return Math.pow(func(t1) - getMediumValue(), 2);
     }
 
@@ -57,5 +50,4 @@ public class TriangularSignal extends Signal {
         this.setVariance(MediumValueORAbsolutMediumValueORMediumPowerORVarianceC(this.getStartTime(), t2, this::funcVar));
         this.setEffectiveValue(EffectiveValueC(this.getStartTime(), t2, this::funcPow));
     }
-
 }
