+++ zad1/src/test/java/tul/ftims/cps/model/GaussianNoiseTest.java
@@ -1,6 +1,8 @@
 package tul.ftims.cps.model;
 
 import org.junit.jupiter.api.Test;
+import tul.ftims.cps.model.manager.Signal;
+import tul.ftims.cps.model.signals.GaussianNoise;
 
 class GaussianNoiseTest {
